requirements.txt
# Core Dependencies
streamlit==1.29.0
python-dotenv==1.0.0

# LLM and AI
euriai
httpx==0.27.0

# Text Splitting
langchain==0.1.0

# Document Processing  
pypdf==3.17.0
python-docx==1.1.0
openpyxl==3.1.2
pandas==2.1.4
pillow==10.1.0
pytesseract==0.3.10

# Vector Store and Embeddings
sentence-transformers==2.3.1
huggingface-hub==0.20.3
tokenizers==0.15.0
transformers==4.36.2
faiss-cpu==1.7.4

# Utilities
numpy==1.26.2
requests==2.31.0

# PDF Export
reportlab==4.0.7
markdown==3.5.1
